Create SMART TRAFFIC LIGHT
To create a smart traffic light system using an Arduino Uno, IR sensors, and LEDs, we will use the two IR sensors to detect the presence of an object. When an object is detected by the first IR sensor, the corresponding LED will turn ON. Once the object passes through the second IR sensor, the LED will turn OFF. The system will use a threshold value to determine whether an object is detected based on the sensor readings.